generalise ron asset loader
