+++ src/util/mod.rs
@@ -1,0 +1,1 @@
+pub mod ron_asset_loader;
