+++ src/util/ron_asset_loader.rs
@@ -1,0 +1,35 @@
+use bevy::{
+    asset::{AssetLoader, LoadContext, io::Reader},
+    prelude::*,
+};
+use serde::de::DeserializeOwned;
+
+#[derive(Default)]
+pub struct RonAssetLoader<T> {
+    _marker: std::marker::PhantomData<T>,
+}
+
+impl<T> AssetLoader for RonAssetLoader<T>
+where
+    T: Asset + DeserializeOwned + Send + Sync + 'static,
+{
+    type Asset = T;
+    type Settings = ();
+    type Error = Box<dyn std::error::Error + Send + Sync>;
+
+    async fn load(
+        &self,
+        reader: &mut dyn Reader,
+        _settings: &(),
+        _load_context: &mut LoadContext<'_>,
+    ) -> Result<Self::Asset, Self::Error> {
+        let mut bytes = Vec::new();
+        reader.read_to_end(&mut bytes).await?;
+        let custom_asset = ron::de::from_bytes::<T>(&bytes)?;
+        Ok(custom_asset)
+    }
+
+    fn extensions(&self) -> &[&str] {
+        &["ron"]
+    }
+}
